exercise-10: updated get items to call db
