+++ todo_app/item.py
@@ -1,10 +1,9 @@
 class Item:
-    def __init__(self, id, name, status, list_id):
+    def __init__(self, id, name, status):
         self.id = id
         self.name = name
         self.status = status
-        self.list_id = list_id
-        
+
     @classmethod
-    def from_trello_card(cls, card, list):
-        return cls(card['id'], card['name'], list['name'], card['idList'])
+    def from_document(cls, doc):
+        return cls(doc['id'], doc['name'], doc['status'])
